Implemented Edit and Delete project

+++ TWC_Backend/TWC_Services/DBService/Services/DBProjectService.cs
@@ -17,20 +17,25 @@ namespace TWC_Services.DBService.Services
         {
             try
             {
-                Project newProject = new Project()
+                var newProject = new Project()
                 {
                     Name = project.Name,
                     MaxMembers = project.MaxMembers,
                     Description = project.Description,
                     Members = new List<ProjectMember>(),
                     Tags = new List<ProjectTag>(),
-
                 };
 
                 await _context.Projects.AddAsync(newProject);
                 await _context.SaveChangesAsync();
 
                 User owner = await _context.Users.FindAsync(project.CreatorId);
+
+                if (owner == null)
+                {
+                    return null;
+                }
+
                 newProject.Members.Add(new ProjectMember()
                 {
                     IsOwner = true,
@@ -43,6 +48,12 @@ namespace TWC_Services.DBService.Services
                 foreach (string tagDTO in project.Tags)
                 {
                     Tag tag = await _context.Tags.SingleOrDefaultAsync(x => x.Name == tagDTO);
+
+                    if(tag == null)
+                    {
+                        return null;
+                    }
+
                     ProjectTag projectTag = new ProjectTag()
                     {
                         TagId = tag.Id,
@@ -67,30 +78,78 @@ namespace TWC_Services.DBService.Services
             }
         }
 
-        public async void DeleteProjectAsync(int id)
+        public async Task DeleteProjectAsync(int id)
         {
-            Project project = await _context.Projects.FindAsync(id);
-            _context.Projects.Remove(project);
-            await _context.SaveChangesAsync();
+            try
+            {
+                Project project = await _context.Projects.FindAsync(id);
 
-            return;
+                if (project == null)
+                {
+                    return;
+                }
+
+                List<ProjectMember> members = await _context.ProjectMembers
+                    .Where(x => x.ProjectId == id).ToListAsync();
+                List<ProjectTag> projectTags = await _context.ProjectTags
+                    .Where(x => x.ProjectId == id).ToListAsync();
+
+                _context.ProjectMembers.RemoveRange(members);
+                _context.ProjectTags.RemoveRange(projectTags);
+                _context.Projects.Remove(project);
+
+                await _context.SaveChangesAsync();
+
+                return;
+            }
+            catch (Exception ex)
+            {
+                throw new Exception($"Exception: {ex}");
+            }
         }
-
-        public async Task<Project> EditProjectAsync(Project project)
+        //Видаляти всіх старих мемберів та теги, проект залишати і використовувати тільки один SaveChangesAsync().
+        public async Task<Project> EditProjectAsync(ProjectEditDTO project)
         {
-            throw new NotImplementedException();
+            try
+            {
+                int projectId = project.Id;
+
+                Project EditProject = await _context.Projects.FindAsync(projectId);
+
+                if(EditProject == null)
+                {
+                    return null;
+                }
+
+                EditProject.Name = project.Name;
+                EditProject.Description = project.Description;
+                EditProject.MaxMembers = project.MaxMembers;
+                EditProject.PictureURL = project.PictureURL;
+
+                await _context.SaveChangesAsync();
+
+                await _context.Entry(EditProject).Collection(x => x.Tags).LoadAsync();
+                await _context.Entry(EditProject).Collection(x => x.Tags).LoadAsync();
+                await _context.Entry(EditProject).Collection(x => x.Members).LoadAsync();
+
+                return EditProject;
+            }
+            catch(Exception ex)
+            {
+                throw new Exception($"Exception: {ex}");
+            }
         }
 
         public async Task<List<Project>> GetAllProjectsAsync()
         {
-            return await _context.Projects.Include(x=>x.Tags).Include(y=> y.Members).ToListAsync();
+            return await _context.Projects.Include(x => x.Tags).Include(y => y.Members).ToListAsync();
         }
 
         public async Task<Project> GetProjectByNameAsync(string name)
         {
             try
             {
-                Project project = await _context.Projects.SingleOrDefaultAsync(x => x.Name.Equals(name));
+                Project project = await _context.Projects.SingleOrDefaultAsync(x => x.Name != null && x.Name.Equals(name));
 
                 return project;
 
@@ -105,14 +164,12 @@ namespace TWC_Services.DBService.Services
         {
             try
             {
-                Project project = await _context.Projects.SingleOrDefaultAsync(x => x.Id == id);
-
+                Project project = await _context.Projects.SingleOrDefaultAsync(x=> x.Id == id);
                 return project;
-
             }
             catch (Exception ex)
             {
-                throw new Exception($"cant found {typeof(Project).Name} by id. Messege error: " + ex.Message);
+                return null;
             }
         }
         private async Task<Project> AddMemberToProjectAsync(User user, Project project, bool isOwner)
